PG Server Additions

+++ index.js
@@ -7,6 +7,7 @@
 const Discord = require('discord.js');
 const client = new Discord.Client();
 const profaneWords = require('./moderation_assets/profane.json');
+const pgservers = require('./moderation_assets/pg_servers.json');
 var prefix = "~";
 
 client.on('ready', () => {
@@ -48,45 +49,52 @@ My command prefix is **~**
   }
 
   // --- PROFANE WORD DETECTION ---
-var msg_pure = message.content.replace(/[^a-zA-Z0-9 ]/g, "").toString().trim();
-for (let i = 0; i < profaneWords.length; i++) {
-  const elem = profaneWords[i];
-  const message_split = msg_pure.split(" ");
-  for (let i = 0; i < message_split.length; i++) {
-    const word = message_split[i];
-    if (word.toLowerCase() == elem) {
-      message.delete();
-      const warning = new Discord.MessageEmbed()
-        .setColor('#FF0000')
-        .setTitle('Profanity Detected!')
-        .setDescription(`Woah, **${message.author.username}**! This is a family friendly server, don't use that language or you may get muted!`)
-        .setFooter('I am still learning! If you feel wrongly judged, please PM one of the server admins and inform them of what you typed.')
-      message.channel.send(warning);
+for (let i = 0; i < pgservers.length; i++) {
+    const elem = pgservers[i];
+    if (message.guild.id == elem) {
+      // PG SERVER do nothing
       return;
+    } else {
+      var msg_pure = message.content.replace(/[^a-zA-Z0-9 ]/g, "").toString().trim();
+      for (let i = 0; i < profaneWords.length; i++) {
+        const elem = profaneWords[i];
+        const message_split = msg_pure.split(" ");
+        for (let i = 0; i < message_split.length; i++) {
+          const word = message_split[i];
+          if (word.toLowerCase() == elem) {
+            message.delete();
+            const warning = new Discord.MessageEmbed()
+              .setColor('#FF0000')
+              .setTitle('Profanity Detected!')
+              .setDescription(`Woah, **${message.author.username}**! This is a family friendly server, don't use that language or you may get muted!`)
+              .setFooter('I am still learning! If you feel wrongly judged, please PM one of the server admins and inform them of what you typed.')
+            message.channel.send(warning);
+            return;
+          }
+          if (word.toLowerCase().startsWith(elem)) {
+            message.delete();
+            const warning = new Discord.MessageEmbed()
+              .setColor('#FF0000')
+              .setTitle('Profanity Detected!')
+              .setDescription(`Woah, **${message.author.username}**! This is a family friendly server, don't use that language or you may get muted!`)
+              .setFooter('I am still learning! If you feel wrongly judged, please PM one of the server admins and inform them of what you typed.')
+            message.channel.send(warning);
+            return;
+          }
+          if (word.toLowerCase().endsWith(elem)) {
+            message.delete();
+            const warning = new Discord.MessageEmbed()
+              .setColor('#FF0000')
+              .setTitle('Profanity Detected!')
+              .setDescription(`Woah, **${message.author.username}**! This is a family friendly server, don't use that language or you may get muted!`)
+              .setFooter('I am still learning! If you feel wrongly judged, please PM one of the server admins and inform them of what you typed.')
+            message.channel.send(warning);
+            return;
+          }
+        }
+      }
     }
-    if (word.toLowerCase().startsWith(elem)) {
-      message.delete();
-      const warning = new Discord.MessageEmbed()
-        .setColor('#FF0000')
-        .setTitle('Profanity Detected!')
-        .setDescription(`Woah, **${message.author.username}**! This is a family friendly server, don't use that language or you may get muted!`)
-        .setFooter('I am still learning! If you feel wrongly judged, please PM one of the server admins and inform them of what you typed.')
-      message.channel.send(warning);
-      return;
-    }
-    if (word.toLowerCase().endsWith(elem)) {
-      message.delete();
-      const warning = new Discord.MessageEmbed()
-        .setColor('#FF0000')
-        .setTitle('Profanity Detected!')
-        .setDescription(`Woah, **${message.author.username}**! This is a family friendly server, don't use that language or you may get muted!`)
-        .setFooter('I am still learning! If you feel wrongly judged, please PM one of the server admins and inform them of what you typed.')
-      message.channel.send(warning);
-      return;
-    }
-  }
 }
-
 });
 
 client.login(process.env.DISCORD_TOKEN);
